+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This changelog's format is based on [keep a changelog 1.0.0](https://keepachange
 - Trace-reading commands now read the trace from the standard input when `-` is specified as the trace filename.
 - Added the `gen-trace` command for generating random traces
 - Added `--parse-permissive` parsing mode for traces
+- `monkey` now performs IPASIR lookups within the current process' own symbols if the specified DSO name is `preloaded`
 
 ### Changed